# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hush Service API
  version: 1.0.0
  description: Single-owner encrypted vault. CLI devices are read-only. All binary
    fields use canonical unpadded base64url. See crypto.md for mandatory AAD,
    algorithm, size, and nonce rules. The Worker validates structure, not
    cryptographic correctness.
servers:
  - url: /api/v1
paths:
  /vault:
    get:
      operationId: getVault
      summary: Read public vault metadata
      security: &a4
        - &a1
          AccessSession: []
        - AccessClientId: []
          AccessClientSecret: []
      parameters: []
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: &a2
                type: string
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultMetadata"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: initializeVault
      summary: Initialize the single vault
      security: &a3
        - *a1
      parameters:
        - name: If-Match
          in: header
          required: true
          description: Quoted vault revision, e.g. "3". Initialization requires "0".
          schema:
            type: string
            pattern: ^"(0|[1-9][0-9]*)"$
        - name: Origin
          in: header
          required: true
          description: Must equal configured ADMIN_ORIGIN.
          schema: *a2
        - name: X-Hush-Request
          in: header
          required: true
          schema:
            const: "1"
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Init"
  /vault/recovery:
    get:
      operationId: getRecovery
      summary: Read master wrapping for owner recovery
      security: *a3
      parameters: []
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recovery"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /vault/snapshot:
    get:
      operationId: getSnapshot
      summary: Read a consistent owner snapshot for rotation
      security: *a3
      parameters: []
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /vault/master-wrap:
    put:
      operationId: changeMasterWrap
      summary: Replace master wrapping with a fresh KDF salt
      security: *a3
      parameters:
        - name: If-Match
          in: header
          required: true
          description: Quoted vault revision, e.g. "3". Initialization requires "0".
          schema:
            type: string
            pattern: ^"(0|[1-9][0-9]*)"$
        - name: Origin
          in: header
          required: true
          description: Must equal configured ADMIN_ORIGIN.
          schema: *a2
        - name: X-Hush-Request
          in: header
          required: true
          schema:
            const: "1"
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChange"
...
components:
  schemas:
    Init:
      type: object
      properties:
        id: &a19
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        name: &a20
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[^\x00-\x1f\x7f]+$
        kdf: &a21
          type: object
          properties:
            algorithm:
              type: string
              const: Argon2id
            version:
              type: number
              const: 19
            salt:
              type: string
              minLength: 22
              maxLength: 22
              pattern: ^[A-Za-z0-9_-]+$
            memory_kib:
              type: number
              const: 65536
            iterations:
              type: number
              const: 3
            parallelism:
              type: number
              const: 1
          required:
            - algorithm
            - version
            - salt
            - memory_kib
            - iterations
            - parallelism
          additionalProperties: false
        master_wrap: &a22
          type: object
          properties:
            format:
              type: number
              const: 1
            algorithm:
              type: string
              const: AES-256-GCM
            nonce:
              type: string
              minLength: 16
              maxLength: 16
              pattern: ^[A-Za-z0-9_-]+$
            ciphertext:
              type: string
              minLength: 64
              maxLength: 64
              pattern: ^[A-Za-z0-9_-]+$
          required:
            - format
            - algorithm
            - nonce
            - ciphertext
          additionalProperties: false
      required:
        - id
        - name
        - kdf
        - master_wrap
      additionalProperties: false
...
    PasswordChange:
      type: object
      properties:
        kdf:
          type: object
          properties:
            algorithm:
              type: string
              const: Argon2id
            version:
              type: number
              const: 19
            salt:
              type: string
              minLength: 22
              maxLength: 22
              pattern: ^[A-Za-z0-9_-]+$
            memory_kib:
              type: number
              const: 65536
            iterations:
              type: number
              const: 3
            parallelism:
              type: number
              const: 1
          required:
            - algorithm
            - version
            - salt
            - memory_kib
            - iterations
            - parallelism
          additionalProperties: false
        master_wrap:
          type: object
          properties:
            format:
              type: number
              const: 1
            algorithm:
              type: string
              const: AES-256-GCM
            nonce:
              type: string
              minLength: 16
              maxLength: 16
              pattern: ^[A-Za-z0-9_-]+$
            ciphertext:
              type: string
              minLength: 64
              maxLength: 64
              pattern: ^[A-Za-z0-9_-]+$
          required:
            - format
            - algorithm
            - nonce
            - ciphertext
          additionalProperties: false
      required:
        - kdf
        - master_wrap
      additionalProperties: false
...
    Secret:
      type: object
      properties:
        name: *a6
        version: *a7
        vek_version: *a8
        envelope: *a9
        id: &a24
          type: string
          format: uuid
        created_at: &a10
          type: string
          format: date-time
        updated_at: *a10
      required:
        - name
        - version
        - vek_version
        - envelope
        - id
        - created_at
        - updated_at
      additionalProperties: false
    Profile:
      type: object
      properties:
        name: *a11
        mappings: *a12
        id:
          type: string
          format: uuid
        created_at: *a10
        updated_at: *a10
      required:
        - name
        - mappings
        - id
        - created_at
        - updated_at
      additionalProperties: false
    Device:
      type: object
      properties:
        id: &a25
          type: string
          format: uuid
        name: *a13
        type: *a14
        public_key: *a15
        access_client_id: *a16
        vek_version: *a17
        wrapped_vek:
          anyOf:
            - *a18
            - type: "null"
        created_at: *a10
        updated_at: *a10
        status: &a26
          enum:
            - active
            - revoked
      required:
        - id
        - name
        - type
        - public_key
        - access_client_id
        - vek_version
        - wrapped_vek
        - created_at
        - updated_at
        - status
      additionalProperties: false
    Vault:
      type: object
      properties:
        id: *a19
        name: *a20
        kdf: *a21
        master_wrap: *a22
        revision: &a23
          type: integer
          minimum: 1
          maximum: 9007199254740990
        vek_version: *a23
        created_at: *a10
        updated_at: *a10
      required:
        - id
        - name
        - kdf
        - master_wrap
        - revision
        - vek_version
        - created_at
        - updated_at
      additionalProperties: false
    VaultMetadata:
      type: object
      properties:
        id: *a19
        name: *a20
        revision: *a23
        vek_version: *a23
        created_at: *a10
        updated_at: *a10
      required:
        - id
        - name
        - revision
        - vek_version
        - created_at
        - updated_at
      additionalProperties: false
...
    Revision:
      type: object
      properties:
        revision: *a23
      required:
        - revision
      additionalProperties: false
    Snapshot:
      type: object
      properties:
        vault:
          $ref: "#/components/schemas/Vault"
        secrets:
          type: array
          items:
            $ref: "#/components/schemas/Secret"
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/Profile"
      required:
        - vault
        - secrets
        - devices
        - profiles
      additionalProperties: false
    Recovery:
      type: object
      properties:
        vault_id: *a2
        revision: *a23
        vek_version: *a23
        kdf: *a21
        master_wrap: *a22
      required:
        - vault_id
        - revision
        - vek_version
        - kdf
        - master_wrap
      additionalProperties: false
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *a2
            request_id: *a2
          required:
            - code
            - request_id
          additionalProperties: false
      required:
        - error
      additionalProperties: false
  securitySchemes:
    AccessSession:
      type: apiKey
      in: cookie
      name: CF_Authorization
      description: Cloudflare Access human session. Edge forwards the signed
        assertion; Worker verifies it and matches OWNER_EMAIL.
    AccessClientId:
      type: apiKey
      in: header
      name: CF-Access-Client-Id
    AccessClientSecret:
      type: apiKey
      in: header
      name: CF-Access-Client-Secret